requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asyncio==3.4.3
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
cmudict==1.0.32
docopt==0.6.2
frozenlist==1.5.0
gevent==24.11.1
greenlet==3.1.1
grequests==0.7.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
multidict==6.1.0
num2words==0.5.14
propcache==0.2.1
pysyllables==1.0.3
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
slack_bolt==1.22.0
slack_sdk==3.34.0
syllables==1.0.9
tracery==0.1.1
urllib3==2.3.0
uvloop==0.21.0
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
